Revert "Feat/staking unlock tutorial debug (#115)"
This reverts commit ffd40900b1516dcead33f4810733496b0ce012c8.
