--- frontend/.tool-versions
@@ -1,1 +1,0 @@
-nodejs nodejs 23.9.0
