+++ frontend/app/components/staking/BackGround.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { useImages } from "../../hooks/useImages";
 import { View } from "react-native";
 import {
@@ -13,27 +12,25 @@ interface BackGroundProps {
   height: number;
 }
 
-export const BackGround: React.FC<BackGroundProps> = React.memo(
-  ({ width, height }) => {
-    const { getImage } = useImages();
+export const BackGround: React.FC<BackGroundProps> = ({ width, height }) => {
+  const { getImage } = useImages();
 
-    return (
-      <View className="absolute w-full h-full">
-        <Canvas style={{ flex: 1 }} className="w-full h-full">
-          <Image
-            image={getImage("background.staking")}
-            fit="fill"
-            x={0}
-            y={-62}
-            width={width}
-            height={height}
-            sampling={{
-              filter: FilterMode.Nearest,
-              mipmap: MipmapMode.Nearest,
-            }}
-          />
-        </Canvas>
-      </View>
-    );
-  },
-);
+  return (
+    <View className="absolute w-full h-full">
+      <Canvas style={{ flex: 1 }} className="w-full h-full">
+        <Image
+          image={getImage("background.staking")}
+          fit="fill"
+          x={0}
+          y={-62}
+          width={width}
+          height={height}
+          sampling={{
+            filter: FilterMode.Nearest,
+            mipmap: MipmapMode.Nearest,
+          }}
+        />
+      </Canvas>
+    </View>
+  );
+};
