+++ frontend/app/components/staking/StakingAction.tsx
@@ -21,53 +21,55 @@ interface StakingActionsProps {
   disabled?: boolean;
 }
 
-export const StakingAction: React.FC<StakingActionsProps> = React.memo(
-  ({ action, label, disabled = false }) => {
-    const { getImage } = useImages();
-    const [size, setSize] = useState({ width: 0, height: 0 });
+export const StakingAction: React.FC<StakingActionsProps> = ({
+  action,
+  label,
+  disabled = false,
+}) => {
+  const { getImage } = useImages();
+  const [size, setSize] = useState({ width: 0, height: 0 });
 
-    const onLayout = (e: LayoutChangeEvent) => {
-      const { width, height } = e.nativeEvent.layout;
-      setSize({ width, height });
-    };
+  const onLayout = (e: LayoutChangeEvent) => {
+    const { width, height } = e.nativeEvent.layout;
+    setSize({ width, height });
+  };
 
-    return (
-      <TouchableOpacity
-        onPress={action}
-        disabled={disabled}
-        style={styles.button}
-        onLayout={onLayout}
-        activeOpacity={0.8}
-      >
-        {size.width > 0 && size.height > 0 && (
-          <Canvas
-            style={[
-              StyleSheet.absoluteFillObject,
-              { width: size.width, height: size.height },
-            ]}
-          >
-            <Image
-              image={getImage("staking.button.bg")}
-              x={0}
-              y={0}
-              width={size.width}
-              height={size.height}
-              fit="fill"
-              sampling={{
-                filter: FilterMode.Nearest,
-                mipmap: MipmapMode.Nearest,
-              }}
-            />
-          </Canvas>
-        )}
+  return (
+    <TouchableOpacity
+      onPress={action}
+      disabled={disabled}
+      style={styles.button}
+      onLayout={onLayout}
+      activeOpacity={0.8}
+    >
+      {size.width > 0 && size.height > 0 && (
+        <Canvas
+          style={[
+            StyleSheet.absoluteFillObject,
+            { width: size.width, height: size.height },
+          ]}
+        >
+          <Image
+            image={getImage("staking.button.bg")}
+            x={0}
+            y={0}
+            width={size.width}
+            height={size.height}
+            fit="fill"
+            sampling={{
+              filter: FilterMode.Nearest,
+              mipmap: MipmapMode.Nearest,
+            }}
+          />
+        </Canvas>
+      )}
 
-        <Text style={styles.label} numberOfLines={1} adjustsFontSizeToFit>
-          {label}
-        </Text>
-      </TouchableOpacity>
-    );
-  },
-);
+      <Text style={styles.label} numberOfLines={1} adjustsFontSizeToFit>
+        {label}
+      </Text>
+    </TouchableOpacity>
+  );
+};
 
 const BORDER_RADIUS = 8;
