+++ frontend/app/components/store/transactionUpgrade/TxDetails.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { View, Text, Dimensions } from "react-native";
 import {
   Canvas,
@@ -6,7 +5,7 @@ import {
   FilterMode,
   MipmapMode,
 } from "@shopify/react-native-skia";
-
+import React from "react";
 import Animated, { FadeInRight } from "react-native-reanimated";
 import { useImages } from "../../../hooks/useImages";
