+++ frontend/app/configs/achievements.json
@@ -158,7 +158,7 @@
     "name": "Proof of Stake!",
     "category": "STORY",
     "image": "achievements.stake",
-    "updateOn": "StakingUnlocked"
+    "updateOn": "StakingPurchased"
   },
   {
     "id": 23,
