+++ frontend/app/hooks/useTutorialLayout.ts
@@ -16,7 +16,7 @@ export function useTutorialLayout(id: TargetId, enabled: boolean = true) {
   const stepTargets = [bubbleTargetId, highlightTargetId];
 
   const measure = useCallback(() => {
-    if (!enabled || !isTutorialActive) {
+    if (!enabled || !isTutorialActive || !stepTargets.includes(id.toString())) {
       return;
     }
     ref.current?.measureInWindow((x, y, width, height) => {
@@ -57,7 +57,7 @@ export function useTutorialLayout(id: TargetId, enabled: boolean = true) {
 
   useEffect(() => {
     if (enabled) scheduleMeasure();
-  }, [enabled, step, scheduleMeasure]);
+  }, [enabled, step]);
 
   return { ref, onLayout };
 }
