+++ frontend/app/navigation/RootNavigator.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import { View } from "react-native";
