+++ frontend/app/observers/TutorialObserver.tsx
@@ -43,7 +43,6 @@ export class TutorialObserver implements Observer {
       this.setVisible(true);
     });
     this.handlers.set("L2Purchased-dapps", () => this.advanceStep());
-    this.handlers.set("SwitchTab-dApps-dapps", () => this.setVisible(false));
     this.handlers.set("SequenceDone-L2Upgrade", () => this.advanceStep());
     this.handlers.set("ProveDone-prover", () => this.advanceStep());
     this.handlers.set("DaDone-dataAvailability", () => this.advanceStep());
@@ -52,11 +51,7 @@ export class TutorialObserver implements Observer {
   }
 
   async onNotify(eventName: EventType, data?: any): Promise<void> {
-    if (
-      eventName === "SwitchPage" ||
-      eventName === "SwitchStore" ||
-      eventName === "SwitchTab"
-    ) {
+    if (eventName === "SwitchPage" || eventName === "SwitchStore") {
       if (data?.name) {
         eventName = eventName + "-" + data.name;
       }
