+++ frontend/app/pages/StakingPage.tsx
@@ -1,20 +1,18 @@
 import { View, Text, SafeAreaView, ScrollView } from "react-native";
-import { useSafeAreaInsets } from "react-native-safe-area-context";
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { useStakingStore } from "../stores/useStakingStore";
 import { useBalanceStore } from "../stores/useBalanceStore";
 import { Dimensions } from "react-native";
-import { useInterval } from "usehooks-ts";
 import { BackGround } from "../components/staking/BackGround";
 import { PageHeader } from "../components/staking/PageHeader";
 import { SectionTitle } from "../components/staking/SectionTitle";
 import { StatsDisplay } from "../components/staking/StatsDisplay";
 import { StakingAction } from "../components/staking/StakingAction";
-import { StakingUnlock } from "../components/staking/StakingUnlock";
 import { AmountField } from "../components/staking/AmountField";
 // import { Canvas, Image, FilterMode, MipmapMode } from "@shopify/react-native-skia";
 // import { useImages } from "../hooks/useImages";
 
+// You may need to define width, height, and getImage if not already imported:
 const { width, height } = Dimensions.get("window");
 
 const SECOND = 1000;
@@ -22,7 +20,6 @@ const BALANCE_PERCENTAGE = [5, 10, 25, 50, 100];
 
 export const StakingPage: React.FC = () => {
   const {
-    stakingUnlocked,
     amountStaked,
     rewards,
     config,
@@ -34,7 +31,6 @@ export const StakingPage: React.FC = () => {
   } = useStakingStore();
   const { tryBuy, updateBalance, balance } = useBalanceStore();
   const [stakeAmount, setStakeAmount] = useState<number>(0);
-  const insets = useSafeAreaInsets();
   // const { getImage } = useImages();
   const balancePercentages = BALANCE_PERCENTAGE;
   interface GetPercentOfParams {
@@ -47,17 +43,15 @@ export const StakingPage: React.FC = () => {
   }
 
   const [now, setNow] = useState(() => Math.floor(Date.now() / 1000));
-
-  useInterval(
-    () => {
-      setNow(Math.floor(Date.now() / 1000));
-    },
-    stakingUnlocked ? SECOND : null,
-  ); // Only run when staking is unlocked
+  useEffect(() => {
+    const id = setInterval(() => setNow(Math.floor(Date.now() / 1000)), SECOND);
+    return () => clearInterval(id);
+  }, []);
 
   // countdown until next validation
   const next = lastValidation + config.slashing_config.due_time;
   const rem = Math.max(next - now, 0);
+  const days = Math.floor(rem / 86400);
   const hours = Math.floor((rem % 86400) / 3600);
   const minutes = Math.floor((rem % 3600) / 60);
   const seconds = rem % 60;
@@ -195,6 +189,7 @@ export const StakingPage: React.FC = () => {
 
           <View className="bg-[#16161d] border border-[#39274E] rounded-b-md p-3 items-center">
             <Text className="font-Pixels text-5xl text-[#fff7ff]">
+              {days.toString().padStart(2, "0")}:
               {hours.toString().padStart(2, "0")}:
               {minutes.toString().padStart(2, "0")}:
               {seconds.toString().padStart(2, "0")}
@@ -204,44 +199,6 @@ export const StakingPage: React.FC = () => {
           <StakingAction action={validateStake} label="VALIDATE" />
         </View>
       </ScrollView>
-
-      {!stakingUnlocked && (
-        <View
-          style={{
-            position: "absolute",
-            top: 0, // inside the content container
-            bottom: 0, // stops right at the top of the tab bar
-            left: 0,
-            right: 0,
-            justifyContent: "center",
-            alignItems: "center",
-            zIndex: 100,
-          }}
-        >
-          <View
-            style={{
-              position: "absolute",
-              top: 0,
-              bottom: 0,
-              left: 0,
-              right: 0,
-              backgroundColor: "rgba(0,0,0,0.8)",
-            }}
-          />
-          <View
-            style={{
-              width: "85%",
-              backgroundColor: "transparent",
-              borderRadius: 10,
-              padding: 20,
-              alignItems: "center",
-              justifyContent: "center",
-            }}
-          >
-            <StakingUnlock />
-          </View>
-        </View>
-      )}
     </SafeAreaView>
   );
 };
