+++ frontend/package.json
@@ -22,7 +22,7 @@
     "@animatereactnative/marquee": "^0.5.2",
     "@avnu/gasless-sdk": "^0.1.8",
     "@expo/vector-icons": "^14.1.0",
-    "@react-native-async-storage/async-storage": "2.1.2",
+    "@react-native-async-storage/async-storage": "^2.1.2",
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/elements": "^2.3.8",
     "@react-navigation/native": "^7.1.6",
@@ -34,11 +34,8 @@
     "@walletconnect/types": "^2.20.3",
     "@walletconnect/universal-provider": "^2.20.3",
     "@walletconnect/web3wallet": "^1.16.1",
-    "ajv": "^8.17.1",
-    "ajv-keywords": "^5.1.0",
     "ethers": "^6.14.3",
     "expo": "53.0.20",
-    "expo-audio": "~0.4.8",
     "expo-blur": "~14.1.5",
     "expo-constants": "~17.1.6",
     "expo-font": "~13.3.1",
@@ -64,17 +61,18 @@
     "react-native-gesture-handler": "~2.24.0",
     "react-native-get-random-values": "^1.11.0",
     "react-native-randombytes": "^3.6.1",
-    "react-native-reanimated": "~3.17.4",
-    "react-native-safe-area-context": "5.4.0",
+    "react-native-reanimated": "^3.17.5",
+    "react-native-safe-area-context": "^5.4.0",
     "react-native-screens": "~4.11.1",
     "react-native-svg": "15.11.2",
     "react-native-web": "~0.20.0",
     "react-native-webview": "13.13.5",
     "starknet": "^7.4.0",
     "tailwindcss": "^3.4.17",
-    "usehooks-ts": "^3.1.1",
     "uuid": "^11.1.0",
-    "zustand": "^5.0.6"
+    "zustand": "^5.0.6",
+    "usehooks-ts": "^3.1.1",
+    "expo-audio": "~0.4.8"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
